requirements.txt
streamlit>=1.28.0
anthropic>=0.25.0
pandas>=2.1.0
python-dotenv>=1.0.0
requests>=2.31.0